refactor: move black_diff.py to formatters/black_formatter.py
