+++ src/darker/formatters/__init__.py
@@ -1,0 +1,1 @@
+"""Built-in code re-formatter plugins."""
